Move `_unpack_provider` function to `contract_utils`
