+++ starknet_py/contract_utils.py
@@ -1,7 +1,9 @@
-from typing import Optional
+from typing import Optional, Tuple, Union
 
 from starknet_py.common import create_casm_class
 from starknet_py.hash.casm_class_hash import compute_casm_class_hash
+from starknet_py.net.account.base_account import BaseAccount
+from starknet_py.net.client import Client
 
 
 def _extract_compiled_class_hash(
@@ -21,3 +23,21 @@ def _extract_compiled_class_hash(
         )
 
     return compiled_class_hash
+
+
+def _unpack_provider(
+    provider: Union[BaseAccount, Client]
+) -> Tuple[Client, Optional[BaseAccount]]:
+    """
+    Get the client and optional account to be used by Contract.
+
+    If provided with Client, returns this Client and None.
+    If provided with BaseAccount, returns underlying Client and the account.
+    """
+    if isinstance(provider, Client):
+        return provider, None
+
+    if isinstance(provider, BaseAccount):
+        return provider.client, provider
+
+    raise ValueError("Argument provider is not of accepted type.")
